[bug] - echeck de demarage : suppression des dependanes jpa
