--- src/test/java/com/bnpparibas/craft/trackntrace/TrackntraceApplicationTests.java
@@ -1,16 +1,0 @@
-package com.bnpparibas.craft.trackntrace;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class TrackntraceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
